pyproject.toml
[project]
name = "painting-split"
version = "0.1.0"
description = "Split multi-canvas or single-canvas images into sections with coordinate metadata."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0",
    "opencv-python-headless>=4.8",
    "Pillow>=10.0",
    "scikit-image>=0.22",
    "flask>=3.0",
    "werkzeug>=3.0",
    "openai>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
